test: move HighIfJSONGenerator.java to test
